PyroCommon is born
