+++ PyroCommon/API/PulloverSpots.cs
@@ -6,7 +6,7 @@ using Rage;
 
 #endregion
 
-namespace SuperEvents.EventFunctions
+namespace PyroCommon.API
 {
     internal class TupleList<T1, T2> : List<Tuple<T1, T2>>
     {
