+++ PyroCommon/App.config
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="utf-8"?><configuration>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="LSPD First Response" publicKeyToken="null" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-0.4.8433.20359" newVersion="0.4.8433.20359" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="RAGENativeUI" publicKeyToken="null" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-1.9.2.0" newVersion="1.9.2.0" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
+</configuration>
