+++ PyroCommon/Main.cs
@@ -1,0 +1,6 @@
+﻿namespace PyroCommon
+{
+    public class Main
+    {
+    }
+}
