+++ PyroCommon/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.CSharp" version="4.7.0" targetFramework="net48" />
+  <package id="PyroSpecial" version="1.0.1" targetFramework="net48" />
+  <package id="RAGENativeUI" version="1.9.2" targetFramework="net48" />
+  <package id="RagePluginHook" version="1.98.0" targetFramework="net48" />
+</packages>
