+++ SuperEvents/Events/Fight.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using LSPD_First_Response.Mod.API;
+using PyroCommon.Events;
 using Rage;
 using Rage.Native;
 using RAGENativeUI;
