+++ SuperEvents/Events/RecklessDriver.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using LSPD_First_Response.Mod.API;
+using PyroCommon.Events;
 using Rage;
 using SuperEvents.EventFunctions;
