+++ SuperPlugins.sln
@@ -8,6 +8,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TurnOffThatEngine", "TurnOf
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeadlyWeapons", "DeadlyWeapons\DeadlyWeapons.csproj", "{6460C304-75FF-46FD-B4E9-EF4E2B8D7AC7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PyroCommon", "PyroCommon\PyroCommon.csproj", "{6E27F96F-AA33-477F-80F4-4245324BE25C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,9 @@ Global
 		{6460C304-75FF-46FD-B4E9-EF4E2B8D7AC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6460C304-75FF-46FD-B4E9-EF4E2B8D7AC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6460C304-75FF-46FD-B4E9-EF4E2B8D7AC7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E27F96F-AA33-477F-80F4-4245324BE25C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E27F96F-AA33-477F-80F4-4245324BE25C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E27F96F-AA33-477F-80F4-4245324BE25C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E27F96F-AA33-477F-80F4-4245324BE25C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
